# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/faturas": {
      "get": {
        "operationId": "FinanceiroController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "aberta",
                "parcial",
                "paga",
                "cancelada"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FaturaResumoDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "financeiro"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/portal/convite/{token}": {
      "get": {
        "operationId": "PortalController_invitePreview",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalInvitePreviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ]
      }
    },
    "/api/portal/convite/aceitar": {
      "post": {
        "operationId": "PortalController_aceitar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalAcceptInviteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalTokensDto"
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ]
      }
    },
    "/api/portal/login": {
      "post": {
        "operationId": "PortalController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalTokensDto"
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ]
      }
    },
    "/api/portal/refresh": {
      "post": {
        "operationId": "PortalController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalRefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalTokensDto"
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ]
      }
    },
    "/api/portal/logout": {
      "post": {
        "operationId": "PortalController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ]
      }
    },
    "/api/portal/me": {
      "get": {
        "operationId": "PortalController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalMeDto"
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/portal/pets": {
      "get": {
        "operationId": "PortalController_pets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortalPetDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/portal/pets/{id}": {
      "get": {
        "operationId": "PortalController_pet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalPetDetalheDto"
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/portal/agendamentos": {
      "get": {
        "operationId": "PortalController_agendamentos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortalAgendamentoDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/portal/faturas": {
      "get": {
        "operationId": "PortalController_faturas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortalFaturaResumoDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/portal/faturas/{id}": {
      "get": {
        "operationId": "PortalController_fatura",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalFaturaDetalheDto"
                }
              }
            }
          }
        },
        "tags": [
          "portal"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/site/config": {
      "get": {
        "operationId": "SiteController_getConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteConfigDto"
                }
              }
            }
          }
        },
        "tags": [
          "site"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "SiteController_updateConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSiteConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteConfigDto"
                }
              }
            }
          }
        },
        "tags": [
          "site"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/site/config/logo/sign-upload": {
      "post": {
        "operationId": "SiteController_signLogo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignLogoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUploadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "site"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/site/config/logo": {
      "post": {
        "operationId": "SiteController_confirmLogo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmLogoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteConfigDto"
                }
              }
            }
          }
        },
        "tags": [
          "site"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/site/solicitacoes": {
      "get": {
        "operationId": "SiteController_listSolicitacoes",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "nova",
                "confirmada",
                "recusada"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolicitacaoDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "site"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/site/solicitacoes/{id}/confirmar": {
      "post": {
        "operationId": "SiteController_confirmar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriagemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitacaoDto"
                }
              }
            }
          }
        },
        "tags": [
          "site"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/site/solicitacoes/{id}/recusar": {
      "post": {
        "operationId": "SiteController_recusar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriagemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitacaoDto"
                }
              }
            }
          }
        },
        "tags": [
          "site"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/site/solicitacoes/{id}/converter": {
      "post": {
        "operationId": "SiteController_converter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConverterResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "site"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "VETAPP API",
    "description": "API do VETAPP — versionada (/api/v1 nas próximas iterações)",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OkDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "SignUploadResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "uploadUrl": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "uploadUrl"
        ]
      },
...
      "FaturaResumoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "responsavelId": {
            "type": "string"
          },
          "responsavelNome": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "totalCentavos": {
            "type": "number"
          },
          "recebidoCentavos": {
            "type": "number",
            "description": "Total já recebido (soma dos recebimentos)"
          },
          "itens": {
            "type": "number"
          },
          "criadaEm": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "responsavelId",
          "responsavelNome",
          "status",
          "totalCentavos",
          "recebidoCentavos",
          "itens",
          "criadaEm"
        ]
      },
...
      "ConverterResultDto": {
        "type": "object",
        "properties": {
          "responsavelId": {
            "type": "string",
            "description": "ID do cliente/responsável criado"
          },
          "solicitacao": {
            "$ref": "#/components/schemas/SolicitacaoDto"
          }
        },
        "required": [
          "responsavelId",
          "solicitacao"
        ]
      },
...
      "PortalInvitePreviewDto": {
        "type": "object",
        "properties": {
          "responsavelNome": {
            "type": "string"
          },
          "clinicaNome": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "responsavelNome",
          "clinicaNome",
          "email"
        ]
      },
      "PortalAcceptInviteDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token do link de convite"
          },
          "password": {
            "type": "string",
            "example": "senha-forte-aqui",
            "minLength": 8
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "PortalTokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "tenantId": {
            "type": "string",
            "description": "Tenant do tutor — o front guarda p/ o próximo login"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "tenantId"
        ]
      },
      "PortalLoginDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "Tenant (clínica) do portal — o front guarda do convite"
          },
          "email": {
            "type": "string",
            "example": "tutor@email.com"
          },
          "password": {
            "type": "string",
            "example": "senha-forte-aqui"
          }
        },
        "required": [
          "tenantId",
          "email",
          "password"
        ]
      },
      "PortalRefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "PortalLogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "PortalMeDto": {
        "type": "object",
        "properties": {
          "responsavelId": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "clinicaNome": {
            "type": "string"
          }
        },
        "required": [
          "responsavelId",
          "nome",
          "email",
          "clinicaNome"
        ]
      },
      "PortalPetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "especie": {
            "type": "string",
            "nullable": true
          },
          "raca": {
            "type": "string",
            "nullable": true
          },
          "sexo": {
            "type": "string",
            "nullable": true
          },
          "castrado": {
            "type": "boolean"
          },
          "nascimento": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "fotoUrl": {
            "type": "string",
            "nullable": true,
            "description": "URL assinada da foto (curta validade)"
          }
        },
        "required": [
          "id",
          "nome",
          "especie",
          "raca",
          "sexo",
          "castrado",
          "nascimento",
          "status",
          "fotoUrl"
        ]
      },
      "PortalVacinaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "descricao",
          "data"
        ]
      },
...
      "PortalPetDetalheDto": {
        "type": "object",
        "properties": {
          "pet": {
            "$ref": "#/components/schemas/PortalPetDto"
          },
          "vacinas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalVacinaDto"
            }
          },
          "historico": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalHistoricoDto"
            }
          }
        },
        "required": [
          "pet",
          "vacinas",
          "historico"
        ]
      },
      "PortalAgendamentoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "petNome": {
            "type": "string",
            "nullable": true
          },
          "tipoNome": {
            "type": "string",
            "nullable": true
          },
          "profissionalNome": {
            "type": "string",
            "nullable": true
          },
          "inicio": {
            "type": "string"
          },
          "fim": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "titulo",
          "petNome",
          "tipoNome",
          "profissionalNome",
          "inicio",
          "fim",
          "status"
        ]
      },
      "PortalFaturaResumoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "totalCentavos": {
            "type": "number"
          },
          "recebidoCentavos": {
            "type": "number"
          },
          "saldoCentavos": {
            "type": "number"
          },
          "criadaEm": {
            "type": "string"
          },
          "notaNumero": {
            "type": "string",
            "nullable": true,
            "description": "Número da NFS-e, se emitida"
          },
          "notaStatus": {
            "type": "string",
            "nullable": true,
            "description": "Status fiscal da fatura"
          }
        },
        "required": [
          "id",
          "status",
          "totalCentavos",
          "recebidoCentavos",
          "saldoCentavos",
          "criadaEm",
          "notaNumero",
          "notaStatus"
        ]
      },
...
      "PortalFaturaDetalheDto": {
        "type": "object",
        "properties": {
          "fatura": {
            "$ref": "#/components/schemas/PortalFaturaResumoDto"
          },
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalFaturaItemDto"
            }
          }
        },
        "required": [
          "fatura",
          "itens"
        ]
      },
...
      "SiteConfigDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "publicado": {
            "type": "boolean"
          },
          "nomeExibicao": {
            "type": "string"
          },
          "sobre": {
            "type": "string"
          },
          "servicos": {
            "type": "string"
          },
          "endereco": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "horario": {
            "type": "string"
          },
          "corPrimaria": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "publicado"
        ]
      },
      "UpdateSiteConfigDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Slug da URL pública (a-z, 0-9, hífen)"
          },
          "publicado": {
            "type": "boolean"
          },
          "nomeExibicao": {
            "type": "string"
          },
          "sobre": {
            "type": "string"
          },
          "servicos": {
            "type": "string",
            "description": "Serviços em destaque, um por linha"
          },
          "endereco": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "horario": {
            "type": "string"
          },
          "corPrimaria": {
            "type": "string",
            "description": "Cor primária em hex (#RRGGBB)"
          }
        }
      },
      "SignLogoDto": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "example": "image/png"
          }
        },
        "required": [
          "contentType"
        ]
      },
      "ConfirmLogoDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          }
        },
        "required": [
          "key"
        ]
      },
      "SolicitacaoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "petNome": {
            "type": "string"
          },
          "servicoDesejado": {
            "type": "string"
          },
          "preferencia": {
            "type": "string"
          },
          "mensagem": {
            "type": "string"
          },
          "origem": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "observacaoInterna": {
            "type": "string"
          },
          "responsavelId": {
            "type": "string",
            "description": "Cliente criado a partir desta solicitação"
          },
          "criadaEm": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome",
          "telefone",
          "status",
          "criadaEm"
        ]
      },
      "TriagemDto": {
        "type": "object",
        "properties": {
          "observacao": {
            "type": "string",
            "description": "Observação interna (opcional)"
          }
        }
      },
...
    }
  }
}